Refactored code to place offer and request in separate components.
